+++ website/core/Footer.js
@@ -36,13 +36,13 @@ class Footer extends React.Component {
                     </a>
                     <div>
                         <h5>Docs</h5>
-                        <a href={this.docUrl('guides/motivation.html', this.props.language)}>
+                        <a href={this.docUrl('guides/motivation', this.props.language)}>
                             Guide
                         </a>
-                        <a href={this.docUrl('examples/basiccrawler.html', this.props.language)}>
+                        <a href={this.docUrl('examples/basiccrawler', this.props.language)}>
                             Examples
                         </a>
-                        <a href={this.docUrl('api/apify.html', this.props.language)}>
+                        <a href={this.docUrl('api/apify', this.props.language)}>
                             Reference
                         </a>
                     </div>